update
- keypad: add comments for understanding
- main: add test parameters

+++ include/keypad/keypad.h
@@ -4,11 +4,6 @@ Created by flo071
 #define keypadPORT  0b00001111
 #define keypadKEYS  0111111111
 
-//#define ROW1 	0b11111110
-//#define ROW2 	0b11111101
-//#define ROW3 	0b11111011
-//#define ROW4 	0b11110111
-
 #define COLUMN1 0b11111110
 #define COLUMN2 0b11111101
 #define COLUMN3 0b11111011
@@ -19,9 +14,4 @@ Created by flo071
 #define ROW3 0b11011111
 #define ROW4 0b11101111
 
-//#define COLUMN4 0b01111111
-//#define COLUMN3	0b10111111
-//#define COLUMN2	0b11011111
-//#define COLUMN1 0b11101111
-
-uint32_t out=0;
+uint16_t out=0;
